pyproject.toml
[tool.poetry]
name = "orion"
version = "0.1.0"
description = "Trading signals platform that identifies option trading opportunities"
authors = ["Your Name <your.email@example.com>"]
readme = "README.md"
packages = [{include = "orion", from = "src"}]

[tool.poetry.dependencies]
python = "^3.9"
asyncio = "*"
aiohttp = "^3.9"
pandas = "^2.1"
numpy = "^1.26"
pydantic = "^2.5"
pydantic-settings = "^2.1"
click = "^8.1"
yfinance = "^0.2"
alpha-vantage = "^2.3"
pandas-ta = "^0.3"
cachetools = "^5.3"
structlog = "^24.1"
tenacity = "^8.2"
pyyaml = "^6.0"
prefect = "^2.14"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4"
pytest-asyncio = "^0.23"
pytest-mock = "^3.12"
pytest-cov = "^4.1"
black = "^23.12"
ruff = "^0.1"
mypy = "^1.8"

[tool.poetry.scripts]
orion = "orion.cli:cli"

[tool.black]
line-length = 100
target-version = ['py311']
include = '\.pyi?$'

[tool.ruff]
line-length = 100
target-version = "py311"
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "I",   # isort
    "B",   # flake8-bugbear
    "C4",  # flake8-comprehensions
    "UP",  # pyupgrade
]
ignore = [
    "E501",  # line too long (handled by black)
    "B008",  # do not perform function calls in argument defaults
]

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
no_implicit_optional = true

[[tool.mypy.overrides]]
module = [
    "yfinance.*",
    "alpha_vantage.*",
    "pandas_ta.*",
    "cachetools.*",
]
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
asyncio_mode = "auto"
markers = [
    "integration: marks tests as integration tests (deselect with '-m \"not integration\"')",
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
]

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
